RDKB-58167: WFO doesn't work with XER10/XLE combo setup
Reason for change: Align data types

Test Procedure:
1. Check if IDM is establishing connection

Risks: Low

Risks: Low
(cherry picked from commit 7a0cc5de50b890a73644e62cc439654f73a7d7f8)

RDKB-58167: WFO doesn't work with XER10/XLE combo setup

Reason for change: Remove packed for structures

Test Procedure:
1. Check if IDM is establishing connection

Risks: Low

Risks: Low

+++ source/InterDeviceManager/Idm_call_back_apis.c
@@ -332,7 +332,7 @@ bool check_device_reachability (char *ip)
 {
     char    cmd[MIN_BUFF] = {0},buf[MIN_BUFF] = {0},out[MAX_BUFF] = {0};
     FILE    *fp = NULL;
-    size_t  total_read = 0;
+    uint32_t total_read = 0;
 
     if (ip == NULL)
     {
@@ -348,7 +348,7 @@ bool check_device_reachability (char *ip)
     }
     memset(out, 0, MAX_BUFF);
     while (fgets(buf, MIN_BUFF, fp) != NULL) {
-        size_t len = strlen(buf);
+        uint32_t len = strlen(buf);
         if (total_read + len >= MAX_BUFF) {
             CcspTraceInfo(("Exceeded buffer size, clipping output\n"));
             break;
