+++ source/InterDeviceManager/Idm_data.c
@@ -233,7 +233,7 @@ ANSC_STATUS IdmMgr_GetFactoryDefaultValue(const char * param_name,char * param_v
 {
 
     FILE * fp = NULL;
-    size_t len = 0;
+    uint32_t len = 0;
     char *line = NULL;
 
     if ((param_name == NULL))
